refactor: Book의 도서관 소장 정보 및 대출 정보를 담고 있는 BookLoanStatePanel 분리
